start of the project
getDealsWeekend.txt holds the date of what the next weekend shall be. 
getDeals.vbs is a vbs script that checks when the next weekend is, and launches a chrome window with skyscanner search for the next weekend (or N weekends from now, specified by a variable) from the "origin" variable.